[core] Extract the Zarr format into heat/core/io/zarr.py
`supports_zarr`, `load_zarr` and `save_zarr` move out of io.py. The
format registers itself and owns its `.zarr` extension constant.

`supports_zarr` is now built on `_config.is_installed`, and the library
is resolved inside the two functions via `require_dependency`, so zarr is
no longer imported at `import heat` time.

Its two hand-rolled `fnmatch` extension checks are replaced by the shared
`utils.sanitize_extension`, so the extension is now recognized the same
way here as in `load`/`save`, which use `splitext`.
